Rename "Edge" to "soma_subscription_fetcher"
The name "Edge" has always confused me -- it sounds like a generic
thing, but it's actually just the code that interfaces NDN with the SOMA
subscriptions server.

The code is also a bit ugly right now, but we'll clean it up later. For
now, just rename it so people don't get caught on the word "Edge" like I
was.