Remove Avahi component from edge-router
It never made sense to me -- why combine it with the part that fetches
from SOMA servers? It also wasn't working yet. If we want to bring back
Avahi, which we eventually want to do -- I think we can make it a
separate daemon, or at least a bit more carefully integrated.