chunks: Make sure subsequent requests go to the fully-qualified name
The protocol here is slightly obtuse, but the goal is to have:

  Request(/manifest.json) -> Data(/manifest.json/$date/$segment(0)) ->
  Request(/manifest.json/$date/$segment(1)) -> Data(/manifest.json/$date/$segment(1))

The name from the data packet, sans segment ID, is known as the
"qualified name". For now, the file producer's resume support is
disabled because we don't save this qualified name, and we need to
ensure that we're requesting segments from the same version.

We also don't yet have a producer that does this redirection. It will
come soon.

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>