usb: Exit on a timeout if there are no interesting mounts available
Almost all of the time, there will not be any interesting mounts
available to advertise NDN data from, so there is no point in the USB
producer running. If a new mount becomes available, the producer will be
re-launched by systemd; so exit after a timeout if no interesting mounts
are around.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T13934