rename endless namespace to ndn:/endlessm
while com.endlessm is a valid name, it doesn't make sense semantically, as 'com.endlessm' represents 2 levels of naming and hence should be /com/endlessm/. but as we have no knowledge of how /com will be handled we can't really use that. strictly speaking and semantically, it would make more sense to use /endless-ndn/ (compared to /ndn/ used by the testbed). but /endlessm will do, and can be linked later on to /com/endlessm once a /com authority exists.

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>