debian: Run all systemd services as ndn-user
This drops their privileges from root a little, and also ensures their
HOME is set the same as for nfd, which means they share the same key
store.

The ndn-user is created in a postinst script, the same as it is in the
nfd package. This is a no-op if the user already exists.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14639