debian: Add missing build-deps and deps
The build-deps were missed because they were already installed on the
development system when testing the packaging. I don’t know why the deps
were missing — dh_python is supposed to detect them automatically and
put them in ${python:Depends}. Something to figure out and fix later.

This includes the addition of python-dbus and gir1.2-notify-0.7 as
dependencies, both of which are deprecated in favour of functionality
which is now in GIO. We should port away from them in future; but they
work well enough for now.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14639