Change the names of shards once again
The last commit didn't quite work -- the names need to be in one HTTP
component for the fetcher to know how to dispatch. Change it so that the
files are flat on disk and we use URI encoding -- just like NDN does in
its URIs. This should export SOMA names on the exact same name.