dbus_service: Fix logging output by setting up a log handler
This ensures that messages outputted by the logging module end up in the
systemd journal, with the identifier
`com.endlessm.EknSubscriptionsDownloader`; they don’t have the
identifier `edd-dbus-service` because systemd does not have a unit file
for them to map the D-Bus well-known name to a systemd unit name. This
can only be fixed once we switch the system to use systemd user sessions
and add unit files for D-Bus session services.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14639