file: Fix calculation of the final segment index
If this method is supposed to return the 0-based index of the final
valid segment in the producer, then it was wrong when the file size was
a multiple of the chunk size, and when the file size was 0.

The method now returns -1 when there is no final segment (i.e. the file
has size 0), and a 0-based index otherwise, addressing a final chunk of
size (0, chunk_size].

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15105