docs: Add documentation infrastructure using Sphinx
There is no API documentation for the moment, but the framework is now
there to start building it as soon as we start adding docstrings.

Note that only a few of the eos_data_distribution.ndn modules are listed
in the index for now. As docstrings are added to other modules, those
modules should be added to the index too.

The documentation is built at package build time, but there is currently
no python-eos-data-distribution-doc package, as I cannot get
dh_sphinxdoc to behave and find the built documentation.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15105