avahi_monitor: route to all NDN nodes only if router not in path
We learned with the Honduras demo that having 30 NDN nodes talking to each
other (relaying interests) could kill a wifi connection. So we implement
this to only route between each other if we can't find a 'master' node. here
the master node is the one that is acting as our gateway

this introduces a new dependency on python-netifaces

Note that we don't use NM because avahi should report routers getting in and
out of scope independantly on networking change. There is a possibility for
a network change to break this setup (like someone switching route manually
from one machine to the other without putting down the iface), but we decide
to ignore that case

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>