file: Fix file creation for interest names starting in a slash
os.path.join() ignores all path components preceding one which starts in
a slash, so we have to ensure the interest’s name doesn’t begin with
one. While we’re at it, add a check to ensure the canonical constructed
path doesn’t escape the directory we’re trying to put it in. This
avoids, for example, being attacked by files named `../../../etc/passwd`
or similar.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15105