chunks: Raise NotImplementedError for abstract methods
There are a couple of abstract methods in this module which derived
classes are meant to implement. Instead of `pass`ing as the
implementation, make it more obvious when something’s wrong by raising a
NotImplementedError.

Ideally, we’d mark these as @abstractmethod and the classes as abstract,
but we can’t do that the normal way (using ABCMeta) because they’re
derived from GObject.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15105