file: Move file creation out of the hot data path
Move it into the slightly-less-hot segment saving path, which only happens
the first time we see a segment. The first time we see *any* segment,
create and open the files.

This fixes a problem in the DirConsumer._on_data method, where the
replacement of the method implementation with its parent wasn’t working,
so _create_files() was being called on every call to _on_data(), and was
re-opening the files on every segment. This resulted in data being lost
if a duplicate segment was received after _on_complete().

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15105