avahi_monitor: start/stop edd-soma-subscriptions-producer
this has to be handled here because it's the place where we know if we are
to be routed or not.

We're also disabling the soma-subscriptions-provider service so that it's
not auto-started under us. We keep the service file around to help
debbugging and allowing manual start (and we use that in this code)

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>