file: Truncate the .part file when opening it
If the file contains junk (either the remains of a previous download, or
just rubbish), we don’t want some of that hanging around after the
download is finished if the .part file was bigger than the final
downloaded file will be. Truncate it to the maximum size we are sure the
final downloaded file will be.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15105