file: Lock files to prevent interference from concurrent downloads
If two independent Consumer instances start downloading the same file,
they would previously overwrite each others’ progress and make a mess.

Add locking to the file.Consumer which prevents this, and causes the
second Consumer instance to try and download a given file to wait until
the first one has finished. Ideally, it would then report success (or
failure) without downloading the file again. However, at the moment
there is no way to verify whether a download was successful or whether
it failed by looking at the downloaded file (for example, no checksum
metadata available), so the second Consumer instance re-downloads the
file. This can be fixed in future.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15105