dbus/base: introduce DbusInstance singleton hack
we create a registery of the different dbus names we're going to listen into, they are actually discriminated by what get_dbusable_name() will generate. for each one of them we centralize requests and keep a registery of the names that we're told about.

this allows us to spawn many many Producers in a process and not kill dbus. As a side effect we now correctly set the dbus object path to find the right producer

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>